feat: `package:args` tab adapter (#4)
* WIP

* Fix option handler boolean flag logic

The `isBool` flag was incorrectly set to `true` when a handler was
provided. This reverses the logic to correctly mark options as boolean
when they have no handler.

Also removes unused files and refactors the tab initialization to use a
simpler constructor without optional parameters.

* Add args adapter for package:args integration

The new adapter automatically generates completions from CommandRunner
and ArgParser configurations, eliminating the need for manual tab
definition when using the args package.

* Mark tab adapters for args package as completed

Update the main README to reflect completion of the tab adapters for the
args package and link to the examples. Also simplify the section header
in the examples README.

* Bump version to 0.0.4

- Fix option handler boolean flag logic
- Add `args` package TAB adapter